+++ partiql-logical-planner/src/lib.rs
@@ -8,7 +8,7 @@ use partiql_ast_passes::name_resolver::NameResolver;
 use partiql_logical as logical;
 use partiql_parser::Parsed;
 
-use partiql_catalog::catalog::{Catalog, PartiqlCatalog};
+use partiql_catalog::catalog::SharedCatalog;
 
 mod builtins;
 mod functions;
@@ -17,11 +17,11 @@ mod lower;
 mod typer;
 
 pub struct LogicalPlanner<'c> {
-    catalog: &'c dyn Catalog,
+    catalog: &'c dyn SharedCatalog,
 }
 
 impl<'c> LogicalPlanner<'c> {
-    pub fn new(catalog: &'c dyn Catalog) -> Self {
+    pub fn new(catalog: &'c dyn SharedCatalog) -> Self {
         LogicalPlanner { catalog }
     }
 
@@ -31,8 +31,7 @@ impl<'c> LogicalPlanner<'c> {
         parsed: &Parsed<'_>,
     ) -> Result<logical::LogicalPlan<logical::BindingsOp>, AstTransformationError> {
         let q = &parsed.ast;
-        let catalog = PartiqlCatalog::default();
-        let mut resolver = NameResolver::new(&catalog);
+        let mut resolver = NameResolver::new(self.catalog);
         let registry = resolver.resolve(q)?;
         let planner = AstToLogical::new(self.catalog, registry);
         planner.lower_query(q)